Fixing the return type for MPProduct quantity getter.

+++ sdk/src/main/java/com/mparticle/MPProduct.java
@@ -99,8 +99,8 @@ public class MPProduct extends HashMap<String, String> {
         return Double.parseDouble(get(UNITPRICE, "0"));
     }
 
-    public int getQuantity(){
-        return Integer.parseInt(get(QUANTITY, "0"));
+    public double getQuantity(){
+        return Double.parseDouble(get(QUANTITY, "0"));
     }
 
     public String getProductCategory(){
